created randomOption component
