+++ src/App.tsx
@@ -5,6 +5,7 @@ import IOption from "./interfaces/IOption";
 import OptionsList from "./components/OptionsList";
 import Header from "./components/Header";
 import { clear } from "console";
+import RandomOption from "./components/RandomOption";
 
 function App() {
     const [options, setOptions] = useState<IOption[]>([]);
@@ -13,15 +14,9 @@ function App() {
     const addOption = (option: IOption) =>
         setOptions([...options, { ...option, id: getCurrentTimestamp() }]);
 
-    const handleRandomOption = () => {
-        const index = Math.floor(Math.random() * options.length);
-        console.log(options[index]);
-    };
 
-    const handleClearOptions = () => {
-      options.splice(0, options.length);
-      console.log(options)
-    }
+
+   
 
     return (
         <>
@@ -33,29 +28,7 @@ function App() {
                     <div className="options-container">
                         <OptionsList options={options} />
                     </div>
-                    <div
-                      style={{
-                        display: "flex",
-                        justifyContent: "center",
-                        gap: "10px",
-                      }}
-                    >
-                        <button
-                            onClick={handleRandomOption}
-                            className={`button ${
-                                options.length === 0 ? "button-empty" : "button"
-                            }`}
-                        >
-                            Elegir!
-                        </button>
-                        <button
-                            
-                            onClick={handleClearOptions}
-                            className="button"
-                        >
-                            Eliminar opciones
-                        </button>
-                    </div>
+                    <RandomOption options={options}/>
                 </div>
             </div>
         </>
